Refactor Inventory table as component, add spinner on cancel
