+++ src/components/WorkOrderDetails.js
@@ -1,6 +1,7 @@
 import React, { Component } from "react";
 import { Link } from "react-router-dom";
 import Button from "./Form/Button";
+import InventoryItemTable from "./WorkOrder/InventoryItemTable";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faClock } from "@fortawesome/free-regular-svg-icons";
 import {
@@ -13,13 +14,11 @@ import {
   faFlagCheckered,
   faMapMarkedAlt,
   faRedo,
-  faTimesCircle,
-  faCheck,
 } from "@fortawesome/free-solid-svg-icons";
 
 import PageTitle from "./Shared/PageTitle";
 import { StyledPageTitle } from "../styles/PageTitle.css";
-import { WorkOrderInventoryStatus } from "../styles/WorkOrderInventoryStatus";
+
 import {
   Accordion,
   AccordionItem,
@@ -50,7 +49,7 @@ class WorkOrderDetail extends Component {
       imagesData: false,
       userInfo: "",
       isSubmitting: false,
-      itemSelectedforCancel: "",
+      atdWorkOrderId: "",
     };
 
     // Split the Details fields in two so we can display them side by side and
@@ -90,6 +89,8 @@ class WorkOrderDetail extends Component {
 
       // Need to retrieve ATD Work Order ID from details in order to req associated inv. items
       const atdWorkOrderId = data.field_1209;
+      // Save atdWorkOrderId for refetching data after inventory updates
+      this.setState({ atdWorkOrderId });
       setTimeout(this.requestInventory, 1500, atdWorkOrderId);
     });
     api
@@ -192,23 +193,6 @@ class WorkOrderDetail extends Component {
         </button>
       </div>
     );
-
-  addWorkOrderStatusClass = status => {
-    // Translate status to classname for CSS styling
-    const statusClassname = status.split(" ")[0].toLowerCase() || null;
-    return !!statusClassname ? statusClassname : "";
-  };
-
-  handleConfirmCancelItemClick = (e, inventoryItemId) => {
-    // Cancel item then remove selected item from state to show item details again
-    const data = { id: inventoryItemId };
-    api
-      .workOrder()
-      .cancelInventoryItem(inventoryItemId, data)
-      .then(() => {
-        this.setState({ itemSelectedforCancel: "" });
-      });
-  };
 
   render() {
     const statusField = this.state.detailsData.field_459;
@@ -313,159 +297,12 @@ class WorkOrderDetail extends Component {
               </h3>
             </AccordionItemTitle>
             <AccordionItemBody>
-              <Button
-                icon={faWrench}
-                text={"New Item"}
-                linkPath={`/work-order/inventory-items/${workOrderId}`}
+              <InventoryItemTable
+                inventoryData={this.state.inventoryData}
+                atdWorkOrderId={this.state.atdWorkOrderId}
+                workOrderId={workOrderId}
+                requestInventory={this.requestInventory}
               />
-              {this.state.inventoryData.length === 0 && <p>No data</p>}
-              {this.state.inventoryData.length > 0 && (
-                <ul className="list-group list-group-flush">
-                  {this.state.inventoryData.map(
-                    (inventory, i) =>
-                      // If an item is selected to cancel, show yes/no confirmation. If not, show item details
-                      this.state.itemSelectedforCancel !== i ? (
-                        <WorkOrderInventoryStatus key={i}>
-                          <li
-                            // Add classname to highlight item name by status
-                            className={`list-group-item d-flex row ${this.addWorkOrderStatusClass(
-                              inventory[workOrderFields.inventory.STATUS]
-                            )}`}
-                            key={i}
-                          >
-                            <div className="col-6">
-                              <div
-                                dangerouslySetInnerHTML={{
-                                  __html:
-                                    inventory[
-                                      workOrderFields.inventory.INVENTORY_ITEM
-                                    ],
-                                }}
-                              />
-                            </div>
-                            <div className="col pt-2">
-                              <Button
-                                icon={faEdit}
-                                text={"Edit"}
-                                linkPath={`/work-order/${workOrderId}/edit-inventory-item/${
-                                  inventory.id
-                                }`}
-                                color={"primary"}
-                              />
-                            </div>
-                            <div className="col pt-2">
-                              {/* Only show cancel button if status is not "Issued" and cancelled is false */}
-                              {inventory[workOrderFields.inventory.STATUS] !==
-                                "Issued" &&
-                                !inventory[
-                                  workOrderFields.inventory.CANCELLED
-                                ] && (
-                                  <div
-                                    className={`btn btn-danger btn-lg`}
-                                    onClick={() =>
-                                      this.setState({
-                                        itemSelectedforCancel: i,
-                                      })
-                                    }
-                                  >
-                                    <FontAwesomeIcon icon={faTimesCircle} />{" "}
-                                    Cancel
-                                  </div>
-                                )}
-                            </div>
-                            <div className="col-12">
-                              {/* Add classname to highlight item attributes by status */}
-                              <div
-                                className={`row ${this.addWorkOrderStatusClass(
-                                  inventory[workOrderFields.inventory.STATUS]
-                                )}`}
-                              >
-                                <div className="col-2">
-                                  <span>Quantity: </span>
-                                  {
-                                    inventory[
-                                      workOrderFields.inventory.QUANTITY
-                                    ]
-                                  }
-                                </div>
-                                <div className="col-2">
-                                  <span>Source: </span>
-                                  {inventory[workOrderFields.inventory.SOURCE]}
-                                </div>
-                                <div className="col-2">
-                                  <span>Issued to: </span>
-                                  {
-                                    inventory[
-                                      workOrderFields.inventory.ISSUED_TO
-                                    ]
-                                  }
-                                </div>
-                                <div className="col-2">
-                                  <span>Comment: </span>
-                                  {inventory[workOrderFields.inventory.COMMENT]}
-                                </div>
-                                <div className="col-2">
-                                  <span>Modified: </span>
-                                  {
-                                    inventory[
-                                      workOrderFields.inventory.MODIFIED
-                                    ]
-                                  }
-                                </div>
-                                <div className="col-2">
-                                  <span>Status: </span>
-                                  {inventory[workOrderFields.inventory.STATUS]}
-                                </div>
-                              </div>
-                            </div>
-                          </li>
-                        </WorkOrderInventoryStatus>
-                      ) : (
-                        <li
-                          className={`list-group-item d-flex row text-center`}
-                          key={i}
-                        >
-                          <div className="col-12 pb-4 lead">
-                            Are you sure you want to cancel this inventory
-                            request?
-                          </div>
-                          <div className="col-6">
-                            <div
-                              className={`btn btn-success btn-lg`}
-                              onClick={e =>
-                                this.handleConfirmCancelItemClick(
-                                  e,
-                                  inventory.id
-                                )
-                              }
-                            >
-                              <FontAwesomeIcon icon={faCheck} /> Yes
-                            </div>
-                          </div>
-                          <div className="col-6">
-                            <div
-                              className={`btn btn-danger btn-lg`}
-                              onClick={() =>
-                                this.setState({ itemSelectedforCancel: "" })
-                              }
-                            >
-                              <FontAwesomeIcon icon={faTimesCircle} /> No
-                            </div>
-                          </div>
-                        </li>
-                      )
-                  )}
-                </ul>
-              )}
-              {!this.state.inventoryData && (
-                <div>
-                  <FontAwesomeIcon
-                    icon={faSpinner}
-                    size="2x"
-                    className="atd-spinner"
-                  />
-                </div>
-              )}
             </AccordionItemBody>
           </AccordionItem>
           <AccordionItem>
